+++ src/main/java/com/intellij/plugins/haxe/ide/generation/HaxeConstructorHandler.java
@@ -22,6 +22,7 @@ package com.intellij.plugins.haxe.ide.generation;
 import com.intellij.openapi.project.Project;
 import com.intellij.plugins.haxe.HaxeBundle;
 import com.intellij.plugins.haxe.lang.psi.HaxeClass;
+import com.intellij.plugins.haxe.lang.psi.HaxeConstructorDeclaration;
 import com.intellij.plugins.haxe.lang.psi.HaxeFieldDeclaration;
 import com.intellij.plugins.haxe.lang.psi.HaxeNamedComponent;
 import com.intellij.plugins.haxe.lang.psi.impl.AbstractHaxeNamedComponent;
@@ -30,8 +31,8 @@ import com.intellij.plugins.haxe.model.HaxeFieldModel;
 import com.intellij.plugins.haxe.model.HaxeMethodModel;
 import com.intellij.plugins.haxe.model.HaxeParameterModel;
 import com.intellij.plugins.haxe.model.type.HaxeTypeResolver;
+import com.intellij.plugins.haxe.model.type.ResultHolder;
 import com.intellij.plugins.haxe.model.type.SpecificTypeReference;
-import com.intellij.plugins.haxe.util.HaxeResolveUtil;
 import com.intellij.plugins.haxe.util.HaxeNamedSubComponentUtil;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.util.PsiTreeUtil;
@@ -41,6 +42,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+
+import static com.intellij.plugins.haxe.ide.inspections.intentions.HaxeIntroduceUtil.findTypesRequiringImportsAndAddToFile;
 
 public class HaxeConstructorHandler extends BaseHaxeGenerateHandler {
   @Override
@@ -54,35 +57,39 @@ public class HaxeConstructorHandler extends BaseHaxeGenerateHandler {
           // Constructor already exists
           return;
         }
+
         if (clazz.isAbstractType()) {
           SpecificTypeReference underlyingType = clazz.getUnderlyingType();
           String parameter  = underlyingType == null || underlyingType.isUnknown() ? "value" :  "value:" +underlyingType.toPresentationString();
           String out = "public function new(" + parameter + ") {\nthis = value;\n}";
            anchor = clazz.getBodyPsi().getFirstChild();
-          doAddMethodsForOne(project, out, anchor);
+          PsiElement psiElement = doAddMethodsForOne(project, out, anchor);
           return;
         }
 
         HaxeMethodModel parentConstructor = clazz.getParentConstructor(null);
 
-        ArrayList<ParamElement> params = new ArrayList<ParamElement>();
-        ArrayList<ParamElement> paramsSuper = new ArrayList<ParamElement>();
-        ArrayList<ParamElement> paramsWrite = new ArrayList<ParamElement>();
+        List<ParamElement> params = new ArrayList<>();
+        List<ParamElement> paramsSuper = new ArrayList<>();
+        List<ParamElement> paramsWrite = new ArrayList<>();
+
         if (parentConstructor != null) {
           for (HaxeParameterModel param : parentConstructor.getParameters()) {
-            ParamElement element = new ParamElement(param.getName(), param.getType().toStringWithoutConstant());
+            ParamElement element = new ParamElement(param.getName(), param.getType().toStringWithoutConstant(), param.isOptional());
             params.add(element);
             paramsSuper.add(element);
           }
         }
         for (HaxeNamedComponent node : elementsToProcess) {
-          ParamElement element = new ParamElement(node.getName(), HaxeTypeResolver
-            .getFieldOrMethodReturnType((AbstractHaxeNamedComponent)node).toStringWithoutConstant());
-          params.add(element);
-          paramsWrite.add(element);
+          if(node instanceof HaxeFieldDeclaration) {
+            ResultHolder fieldOrMethodReturnType = HaxeTypeResolver.getFieldOrMethodReturnType((AbstractHaxeNamedComponent) node);
+            String stringWithoutConstant = fieldOrMethodReturnType.toStringWithoutConstant();
+            ParamElement element = new ParamElement(node.getName(), stringWithoutConstant, false);
+            params.add(element);
+            paramsWrite.add(element);
+          }
         }
 
-        System.out.println(parentConstructor);
 
 
         boolean first = true;
@@ -95,11 +102,18 @@ public class HaxeConstructorHandler extends BaseHaxeGenerateHandler {
         }
 
         String out = createClassConstrcutorString( params, first, parentConstructor, paramsSuper, paramsWrite);
-        doAddMethodsForOne(project, out, anchor);
-        //super.processElements(project, elementsToProcess);
+        PsiElement psiElement = doAddMethodsForOne(project, out, anchor);
+
+        if (psiElement instanceof HaxeConstructorDeclaration constructorDeclaration && parentConstructor != null) {
+          HaxeMethodModel newConstructor = constructorDeclaration.getModel();
+          List<HaxeParameterModel> parameters = newConstructor.getParameters();
+          List<ResultHolder> knownParamTypes = parentConstructor.getParameters().stream().map(HaxeParameterModel::getType).toList();
+          findTypesRequiringImportsAndAddToFile(parameters, knownParamTypes, null, null, psiElement.getContainingFile());
+        }
+
       }
 
-      private static @NotNull String createClassConstrcutorString(ArrayList<ParamElement> params, boolean first, HaxeMethodModel parentConstructor, ArrayList<ParamElement> paramsSuper, ArrayList<ParamElement> paramsWrite) {
+      private static @NotNull String createClassConstrcutorString(List<ParamElement> params, boolean first, HaxeMethodModel parentConstructor, List<ParamElement> paramsSuper, List<ParamElement> paramsWrite) {
         String out = "public function new(";
         for (ParamElement param : params) {
           if (!first) {
@@ -107,6 +121,7 @@ public class HaxeConstructorHandler extends BaseHaxeGenerateHandler {
           } else {
             first = false;
           }
+          out += param.optional ? "?" : "";
           out += param.name;
           if(!param.type.equals("unknown")) {
             out += ":";
@@ -149,15 +164,16 @@ public class HaxeConstructorHandler extends BaseHaxeGenerateHandler {
   @Override
   void collectCandidates(HaxeClass haxeClass, List<HaxeNamedComponent> candidates) {
     final List<HaxeNamedComponent> subComponents = HaxeNamedSubComponentUtil.getNamedSubComponentsFromClassType(haxeClass);
-    final Map<String, HaxeNamedComponent> componentMap = HaxeNamedSubComponentUtil.namedComponentToMap(subComponents);
 
     for (HaxeNamedComponent haxeNamedComponent : subComponents) {
-      if (!(haxeNamedComponent instanceof HaxeFieldDeclaration)) continue;
-      if (haxeNamedComponent.isStatic()) continue;
-
-      //if (!myStrategy.accept(haxeNamedComponent.getName(), componentMap.keySet())) continue;
-
-      candidates.add(haxeNamedComponent);
+      if (haxeNamedComponent instanceof HaxeFieldDeclaration fieldDeclaration) {
+        if (haxeNamedComponent.isStatic()) continue;
+        if(fieldDeclaration.getModel() instanceof  HaxeFieldModel fieldModel) {
+          if(fieldModel.isFinal() && fieldModel.hasInitializer()) continue;
+          if(fieldModel.isInline()) continue;
+          candidates.add(haxeNamedComponent);
+        }
+      }
     }
   }
 }
@@ -165,9 +181,11 @@ public class HaxeConstructorHandler extends BaseHaxeGenerateHandler {
 class ParamElement {
   public String name;
   public String type;
+  public boolean optional;
 
-  public ParamElement(String name, String type) {
+  public ParamElement(String name, String type, boolean optional) {
     this.name = name;
     this.type = type;
+    this.optional = optional;
   }
 }